Add files via upload
This is a randomized analysis of 9 students in a learning community, analyzing the Body Mass Index and getting to know whose B.M.I is not okay. In the data, there is a an outlier. The regression analysis was computed, using the Height alongside the weight to determine our coefficient. In the data, there is an outlier, whose weight affected the coefficient. Hence, it gave us a weak positive coefficient.